requirements.txt
Django==2.2.10
django12factor==1.3
djangorestframework==3.12.4
drf-spectacular==0.15.1
ipython==7.23.0
model-bakery==1.3.1
